数据分析活/requirements.txt
# 基础依赖
pandas==2.0.3
numpy==1.24.3
python-dotenv==1.0.0
pyyaml==6.0.1

# 数据库
mysql-connector-python==8.1.0
redis==4.6.0
sqlalchemy==2.0.19

# 大数据组件
pyspark==3.4.1
apache-flink==1.17.1
kafka-python==2.0.2
hdfs==2.7.0

# Web服务
flask==2.3.3
flask-restful==0.3.10
gunicorn==21.2.0

# 数据分析和机器学习
scikit-learn==1.3.0
prophet==1.1.4
tensorflow==2.13.0

# 监控和日志
prometheus-client==0.17.1
python-json-logger==2.0.7

# 测试
pytest==7.4.0
pytest-cov==4.1.0

# 开发工具
jupyter==1.0.0
black==23.7.0
flake8==6.1.0 